pyproject.toml
[project]
name = "kg-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "fastmcp>=0.3.0",
    "langchain>=1.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-neo4j>=0.6.0",
    "langchain-openai>=1.1.6",
    "neo4j>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.40.0",
]
